requirements/requirements_dev.txt
jupyterlab
pandas
numpy
scikit-learn
matplotlib
seaborn
torch
transformers
pyarrow
janome
wordcloud
japanize-matplotlib
sentence-transformers

# Gemini利用用（新規追加）
google-genai
python-dotenv
google-generativeai

# 可視化ダッシュボード
streamlit